-Modifier la classe Offre & ServiceOffre (Ajouter l'attribut date expiration)
-ajouter l'interface domaine & controller
-Affecter l'ajout d'un offre 

-ajouter l'interface offre & controller
-implementer la methode d'ajout d'un offre

-Ajout des interface AjoutCandidature et DetailsCandidature
-Implementation de la methode qui ajoute et affiche une candidature

-Crud Gerer Domaine
-Crud Gerer Offre

-crud Candidature avec interface Graphique

+++ src/com/esprit/controller/CrudOffreController.java
@@ -26,13 +26,16 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.ChoiceBox;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableColumn.CellEditEvent;
 import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.control.cell.TextFieldTableCell;
 import javafx.scene.input.MouseEvent;
 import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
@@ -47,12 +50,10 @@ public class CrudOffreController implements Initializable {
 
     @FXML
     private TextField txtTitre;
-
     @FXML
     private DatePicker datePicker;
     @FXML
     private TextArea txtDescription;
-
     @FXML
     private Button btnAdd;
     @FXML
@@ -70,7 +71,7 @@ public class CrudOffreController implements Initializable {
     @FXML
     private TableColumn<OffreView, Date> dateExpCol;
     @FXML
-    private ChoiceBox<String> ChoiseBoxDomaine;
+    private ComboBox<String> ChoiseBoxDomaine;
     @FXML
     private TableColumn<OffreView, String> NomDomaineCol;
     @FXML
@@ -78,7 +79,7 @@ public class CrudOffreController implements Initializable {
     @FXML
     private TableColumn<OffreView, String> NomEntCol;
     
-    
+    ServiceOffre sf = new ServiceOffre();
     /**
      * Initializes the controller class.
      * @param url
@@ -88,20 +89,36 @@ public class CrudOffreController implements Initializable {
     public void initialize(URL url, ResourceBundle rb) {
         table();
         ServiceDomaine sd = new ServiceDomaine();
+       
         ChoiseBoxDomaine.getItems().addAll(sd.getDomainesName());
+        
+        titreCol.setCellFactory(TextFieldTableCell.forTableColumn());
+        titreCol.setOnEditCommit(this::editTitre);
+        
+        descCol.setCellFactory(TextFieldTableCell.forTableColumn());
+        descCol.setOnEditCommit(this::editDesc);
+        
+//        datePubCol.setCellFactory(DatePickerTableCell.forTableColumn());
+//        datePubCol.setOnEditCommit(this::editDatePub);
+        
     }    
 
-    @FXML
-    private void selectDate(ActionEvent event) {
-    }
+
 
     @FXML
     private void add(ActionEvent event) {
-        ServiceOffre sf = new ServiceOffre();
+        
         ServiceDomaine sd = new ServiceDomaine();
-        sf.ajouter(new Offre(txtTitre.getText(), txtDescription.getText(),sd.getIdDomaineByName(ChoiseBoxDomaine.getValue()),2, Date.valueOf(datePicker.getValue())));
-        JOptionPane.showMessageDialog(null, "offre ajouter !");
-        table();
+        
+        
+        int result = JOptionPane.showConfirmDialog(null, "vouler vous ajouter ce offre ?","Confirmation",JOptionPane.YES_NO_OPTION);
+            if (result == JOptionPane.YES_OPTION){
+                sf.ajouter(new Offre(txtTitre.getText(), txtDescription.getText(),sd.getIdDomaineByName(ChoiseBoxDomaine.getValue()),2, Date.valueOf(datePicker.getValue())));
+                JOptionPane.showMessageDialog(null, "offre ajouter !");
+                table();
+            }else{
+                return ;
+            }
     }
     
     
@@ -111,7 +128,6 @@ public class CrudOffreController implements Initializable {
     
     @FXML
     private void update(ActionEvent event) throws IOException {
-        
         index = tableOffre.getSelectionModel().getSelectedIndex();
         if (index <= -1) {
             JOptionPane.showMessageDialog(null,"il faut selectionner une ligne !");
@@ -125,13 +141,14 @@ public class CrudOffreController implements Initializable {
             stage.setScene(scene);
             stage.show();
             
-            
+            uoc.setIdOffre(tableOffre.getItems().get(index).getId_offre());
             uoc.setTxtDescription(descCol.getCellData(index));
             uoc.setTxtTitre(titreCol.getCellData(index));
             uoc.setChoiseBoxDomaine();
-            uoc.setComboBoxDomaine();
+            uoc.setIDEntreprise(tableOffre.getItems().get(index).getId_entreprise());
             uoc.setDatePickerPub(Date.valueOf(datePubCol.getCellData(index).toString()));
             uoc.setDatePickerExp(Date.valueOf(dateExpCol.getCellData(index).toString()));
+            
             uoc.setOffreController(this);
             uoc.setStage(stage);
             
@@ -145,11 +162,26 @@ public class CrudOffreController implements Initializable {
     private void delete(ActionEvent event) {
         int index;
         index = tableOffre.getSelectionModel().getSelectedIndex();
-        int id = Integer.parseInt(String.valueOf(tableOffre.getItems().get(index).getId_offre()));
         ServiceOffre sf = new ServiceOffre();
-        sf.supprimerParId(id);
-        JOptionPane.showMessageDialog(null, "offre supprimer !");
-        table();
+        if(index < 0){
+            JOptionPane.showMessageDialog(null, "Selectionner un offre !");
+        }else{
+            int id = Integer.parseInt(String.valueOf(tableOffre.getItems().get(index).getId_offre())); 
+            int result = JOptionPane.showConfirmDialog(null, "vouler vous ajouter ce offre ?","Confirmation",JOptionPane.YES_NO_OPTION);
+            if (result == JOptionPane.YES_OPTION){
+                
+              sf.supprimerParId(id);
+             JOptionPane.showMessageDialog(null, "offre supprimer !");
+                
+                table();
+               
+            }else{
+                return ;
+            }
+        }
+        
+        
+        
     }
     
     public void table(){
@@ -170,35 +202,66 @@ public class CrudOffreController implements Initializable {
 
     @FXML
     private void AfficherOffre(ActionEvent event) throws IOException {
-//        index = tableOffre.getSelectionModel().getSelectedIndex();
-//        if (index <= -1) {
-//            JOptionPane.showMessageDialog(null,"il faut selectionner une ligne !");
-//        }else{
-//            ServiceDomaine sd = new ServiceDomaine();
+        index = tableOffre.getSelectionModel().getSelectedIndex();
+        if (index <= -1) {
+            JOptionPane.showMessageDialog(null,"il faut selectionner une ligne !");
+        }else{
+            
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("../gui/OffreDetails.fxml"));
+            Parent root = loader.load();
+            Scene scene = new Scene(root);
+            Stage stage = new Stage();
+            stage.setScene(scene);
 //            
-//            FXMLLoader loader = new FXMLLoader(getClass().getResource("OffreDetails.fxml"));
-//            Parent root = loader.load();
-//            Scene scene = new Scene(root);
-//            Stage stage = new Stage();
-//            stage.setScene(scene);
+            OffreDetailsController odc = loader.getController();
+            odc.setLNomEntreprise(NomEntCol.getCellData(index).toString());
+            odc.setLDesc(descCol.getCellData(index).toString());
+            odc.setLTitle(titreCol.getCellData(index).toString());
+            odc.setLNomDomaine(NomDomaineCol.getCellData(index));
+            odc.setLDatePub(datePubCol.getCellData(index).toString());
+            odc.setLDateExp(dateExpCol.getCellData(index).toString());
+            odc.setIdoffre(tableOffre.getItems().get(index).getId_offre());
+            odc.setStage(stage);
 //            
-//            OffreDetailsController odc = loader.getController();
-//            odc.setLNomEntreprise(idCol.getCellData(index).toString());
-//            odc.setLDesc(descCol.getCellData(index).toString());
-//            odc.setLTitle(titreCol.getCellData(index).toString());
-//            odc.setLNomDomaine(sd.getNameDomaineById(NomDomaineCol.getCellData(index)));
-//            odc.setLDatePub(datePubCol.getCellData(index).toString());
-//            odc.setLDateExp(dateExpCol.getCellData(index).toString());
-//            
-//            
-//            
-//            
-//            stage.show();
-//           
-//        }
+            
+            
+            stage.show();
+           
+        }
+
+
         
     }
-    
+      
+    @FXML
+    private void editTitre(CellEditEvent event) {
+        int index = tableOffre.getSelectionModel().getSelectedIndex();
+        int id = tableOffre.getItems().get(index).getId_offre();
+        Offre f = sf.chercherOffreByID(id);
+        f.setTitre(event.getNewValue().toString());
+        sf.modifier(f);
+        table();
+    }
+
+    @FXML
+    private void editDesc(CellEditEvent event) {
+        int index = tableOffre.getSelectionModel().getSelectedIndex();
+        int id = tableOffre.getItems().get(index).getId_offre();
+        Offre f = sf.chercherOffreByID(id);
+        f.setDescription(event.getNewValue().toString());
+        sf.modifier(f);
+        table();
+    }
+
+    @FXML
+    private void editDatePub(CellEditEvent event) {
+//        int index = tableOffre.getSelectionModel().getSelectedIndex();
+//        int id = tableOffre.getItems().get(index).getId_offre();
+//        Offre f = sf.chercherOffreByID(id);
+//        f.setDate_offre((Date) event.getNewValue());
+//        sf.modifier(f);
+//        table();
+    }
